requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.2.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.1
multidict==6.1.0
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
